Deduplicate set_mpp and get_raw_faulting_instr
Both instructions performs exactly the same logic in userspace.rs and metal.rs and are not architecture dependent. This commit removes the smell
